WebScraping with Python and BeautifulSoup: Part 1 of 3
In today’s demo, I am going to teach you about the basics of using WebScraping with Python and BeautifulSoup. You’re going to see the objects that comprise Beautiful Soup, and how to work with them. 

In a follow-up demo, I'm going to teach you to work with parsed data in BeautifulSoup. In the second follow-up, you're going to learn how to scrape a webpage and save your results to a working directory on your machine.